'Resolve some bugs and cleaned a little a bit'
