+++ conf/models/segmentation.yaml
@@ -22,6 +22,7 @@ models:
             nn: []
             dropout: 0.5
 
+    # RandLA-Net: Efficient Semantic Segmentation of Large-Scale Point Clouds (https://arxiv.org/pdf/1911.11236.pdf)
     RandLANet:
         name: ranldanet_res
         down_conv:
@@ -42,30 +43,10 @@ models:
             up_k: [1, 1, 1]
             skip: True
         mlp_cls:
-            nn: [64, 64, 64, 64]
+            nn: [64, 64, 64, 64, 64]
             dropout: 0.5
-        # RandLANet:
-        #     name: RandLANet
-        #     down_conv:
-        #         module_name: RandLANet
-        #         ratio: [0.25, 0.25, 0.25]
-        #         k: [16, 16, 16]
-        #         point_pos_nn: [[10, 8, 3], [10, 8, 16], [10, 16, 32]]
-        #         attention_nn: [[6, 8, 6], [32, 64, 32], [64, 128, 64]]
-        #         down_conv_nn: [[6, 8, 16], [32, 64, 32], [64, 128, 128]]
-        #     innermost:
-        #         module_name: GlobalBaseModule
-        #         aggr: max
-        #         nn: [131, 128]
-        #     up_conv:
-        #         module_name: FPModule
-        #         up_conv_nn: [[256, 128], [160, 64], [80, 64], [64, 64]] 
-        #         up_k: [1, 1, 1, 1]
-        #         skip: True
-        #     mlp_cls:
-        #         nn: [64, 64, 64, 64]
-        #         dropout: 0.5
 
+    # Relation-Shape Convolutional Neural Network for Point Cloud Analysis (https://arxiv.org/abs/1904.07601)
     RSConv:
         name: RSConv
         down_conv:
@@ -89,7 +70,7 @@ models:
             nn: [64, 64, 64, 64]
             dropout: 0.5
 
-    # This part describes a KPConvModel model configuration
+    # KPConv: Flexible and Deformable Convolution for Point Clouds (https://arxiv.org/abs/1904.08889)
     KPConv:
         name: KPConv
         down_conv:
@@ -112,6 +93,7 @@ models:
             nn: [64, 64, 64, 64, 64]
             dropout: 0.5
 
+    # PointNet++: Deep Hierarchical Feature Learning on Point Sets in a Metric Space (https://arxiv.org/abs/1706.02413)
     pointnet2:
         name: pointnet2
         down_conv:
