+++ models/core_modules.py
@@ -121,101 +121,3 @@ class GlobalBaseModule(torch.nn.Module):
         batch = torch.arange(x.size(0), device=batch.device)
         data = (x, pos, batch)
         return data
-
-
-
-class BaseSampler(ABC):
-
-    def __init__(self, ratio):
-        self.ratio = ratio
-
-    @abstractmethod
-    def sample(self, pos, batch):
-        pass
-
-class FPSSampler(BaseSampler):
-
-    def __init__(self, ratio):
-        super(FPSSampler, self).__init__(ratio)
-
-    def sample(self, pos, batch):
-        return fps(pos, batch, ratio=self.ratio)
-
-class RandomSampler(BaseSampler):
-
-    def __init__(self, ratio):
-        super(RandomSampler, self).__init__(ratio)
-
-    def sample(self, pos, batch):
-        idx = torch.randint(0, pos.shape[0], (math.floor(pos.shape[0]*self.ratio),))
-        return idx
-
-
-
-class BaseNeighbourFinder(ABC):
-
-    @abstractmethod
-    def find_neighbours(self, x, y, batch_x, batch_y):
-        pass
-
-class RadiusNeighbourFinder(BaseNeighbourFinder):
-
-    def __init__(self, radius, max_num_neighbors=64):
-        self.radius = radius 
-        self.max_num_neighbors = max_num_neighbors
-
-    def find_neighbours(self, x, y, batch_x, batch_y):  
-        return radius(x, y, self.radius, batch_x, batch_y, max_num_neighbors=self.max_num_neighbors)
-
-class KNNNeighbourFinder(BaseNeighbourFinder):
-
-    def __init__(self, k):
-        self.k = k 
-
-    def find_neighbours(self, x, y, batch_x, batch_y):
-        return knn(x, y, self.k, batch_x, batch_y)      
-
-# class Sampler():
-
-#     def __init__(self, sampling_strategy = 'fps', neighbour_strategy = 'radius', ratio = None, opts = None):
-
-#         if sampling_strategy == 'fps':
-#             self.sampling_func = partial(fps, ratio=ratio)
-
-#         elif sampling_strategy == 'random':
-
-#             def random_sample(pos, batch):
-#                 idx = torch.randint(0, pos.shape[0], (math.floor(pos.shape[0]*ratio),))
-#                 return idx
-
-#             self.sampling_func = random_sample
-#         else:
-#             raise ValueError("Unrecognised sampling strategy: ", sampling_strategy)
-
-#         if neighbour_strategy == 'radius':
-
-#             def radius_wrapper(x, y, batch_x, batch_y):
-#                 return radius(x, y, opts['radius'], batch_x, batch_y, max_num_neighbors=opts['max_num_neighbours'])
-
-#             self.neighbour_func = radius_wrapper
-#         elif neighbour_strategy == 'knn':
-
-#             def knn_wrapper(x, y, batch_x, batch_y):
-#                 return knn(x, y, opts['k'], batch_x, batch_y)
-
-#             self.neighbour_func = knn_wrapper
-#         else:
-#             raise ValueError("Unrecognised neighbour strategy: ", neighbour_strategy)
-
-#     def build(self):
-
-#         def sample_func(pos, batch):
-#             idx = self.sampling_func(pos, batch)
-#             row, col = self.neighbour_func(pos, pos[idx], batch, batch[idx])
-#             edge_index = torch.stack([col, row], dim=0)
-#             return edge_index
-
-#         return sample_func
-            
-
-
